+++ tests/testthat/test-fit.R
@@ -64,3 +64,90 @@ test_that("cannot fit without a fit stage", {
   expect_error(fit(workflow, mtcars), "must have a model")
 })
 
+# ------------------------------------------------------------------------------
+# .fit_pre()
+
+test_that("`.fit_pre()` updates a formula blueprint according to parsnip's encoding info", {
+  workflow <- workflow()
+  workflow <- add_formula(workflow, Sepal.Length ~ .)
+
+  mod <- parsnip::rand_forest()
+  mod <- parsnip::set_engine(mod, "ranger")
+  mod <- parsnip::set_mode(mod, "regression")
+  workflow <- add_model(workflow, mod)
+
+  result <- .fit_pre(workflow, iris)
+
+  # ranger sets `indicators = 'none'`, so `Species` is not expanded
+  expected <- "Species"
+  expect_true(expected %in% names(result$pre$mold$predictors))
+  expect_identical(result$pre$actions$formula$blueprint$indicators, "none")
+
+  mod <- parsnip::boost_tree(trees = 5)
+  mod <- parsnip::set_engine(mod, "xgboost")
+  mod <- parsnip::set_mode(mod, "regression")
+  workflow <- update_model(workflow, mod)
+
+  result <- .fit_pre(workflow, iris)
+
+  # xgboost sets `indicators = 'one_hot'`, so `Species` is expanded to three values
+  expected <- c("Speciessetosa", "Speciesversicolor", "Speciesvirginica")
+  expect_true(all(expected %in% names(result$pre$mold$predictors)))
+  expect_identical(result$pre$actions$formula$blueprint$indicators, "one_hot")
+})
+
+test_that("`.fit_pre()` ignores parsnip's encoding info with recipes", {
+  mod <- parsnip::rand_forest()
+  mod <- parsnip::set_engine(mod, "ranger")
+  mod <- parsnip::set_mode(mod, "regression")
+  rec <- recipes::recipe(Sepal.Length ~ ., iris)
+
+  workflow <- workflow()
+  workflow <- add_recipe(workflow, rec)
+  workflow <- add_model(workflow, mod)
+
+  result <- .fit_pre(workflow, iris)
+
+  # recipe preprocessing won't auto-expand factors
+  expect_true("Species" %in% names(result$pre$mold$predictors))
+  expect_false("indicators" %in% names(result$pre$actions$recipe$blueprint))
+})
+
+test_that("`.fit_pre()` doesn't modify user supplied formula blueprint", {
+  mod <- parsnip::rand_forest()
+  mod <- parsnip::set_engine(mod, "ranger")
+  mod <- parsnip::set_mode(mod, "regression")
+
+  # request `indicators` to be used, even though parsnip's info on ranger
+  # says not to make them.
+  blueprint <- hardhat::default_formula_blueprint(indicators = "traditional")
+
+  workflow <- workflow()
+  workflow <- add_formula(workflow, Sepal.Length ~ ., blueprint = blueprint)
+  workflow <- add_model(workflow, mod)
+
+  result <- .fit_pre(workflow, iris)
+
+  expected <- c("Speciessetosa", "Speciesversicolor", "Speciesvirginica")
+  expect_true(all(expected %in% names(result$pre$mold$predictors)))
+  expect_identical(result$pre$actions$formula$blueprint, blueprint)
+})
+
+test_that("`.fit_pre()` doesn't modify user supplied recipe blueprint", {
+  mod <- parsnip::rand_forest()
+  mod <- parsnip::set_engine(mod, "ranger")
+  mod <- parsnip::set_mode(mod, "regression")
+  rec <- recipes::recipe(Sepal.Length ~ ., iris)
+
+  blueprint <- hardhat::default_recipe_blueprint(allow_novel_levels = TRUE)
+
+  workflow <- workflow()
+  workflow <- add_recipe(workflow, rec, blueprint = blueprint)
+  workflow <- add_model(workflow, mod)
+
+  result <- .fit_pre(workflow, iris)
+
+  expect_true("Species" %in% names(result$pre$mold$predictors))
+  expect_identical(result$pre$actions$recipe$blueprint, blueprint)
+})
+